MIDI-Out: removed `if Changed(ch)` to fix veloc, bend, aft, and cc conversion when routing internally. added send delay for pitch bend. rearranged some stuff and updated CommitSettings function mapping for channel B. cleaned up code style for consistency, and fixed some datatypes
TO DO:
fix midi output message log
remove unneccessary NoteOff spam when i1 is not in Note mode.
test CaptainMidi because I changed PB on that too.
purge Changed() from BootsNCat and maybe Trending ?

+++ software/src/HSIOFrame.h
@@ -452,7 +452,7 @@ struct MIDIFrame {
         MIDI1.sendAfterTouch(val, midi_ch + 1);
 #endif
     }
-    void SendPitchBend(const uint8_t midi_ch, uint16_t bend) {
+    void SendPitchBend(const uint8_t midi_ch, int bend) {
         usbMIDI.sendPitchBend(bend, midi_ch + 1);
 #ifdef ARDUINO_TEENSY41
         usbHostMIDI.sendPitchBend(bend, midi_ch + 1);
